New payment management page
New page and security improved when post the api.

+++ src/app/api/v1/del-tenant/route.ts
@@ -1,10 +1,16 @@
 
+import { getSession } from '@/lib/auth';
 import { firestoreService } from '@/lib/services/firestore-service';
 import { Tenant } from '@/types/tenant';
 import { NextResponse } from 'next/server';
 
 export async function POST(req: Request) {
   const tenant:Tenant = await req.json();
+
+  // Check auth
+  if(!getSession()){
+    return NextResponse.json({ success: false, message: "User not authenticated" });
+  }
 
   try {
     await firestoreService.deleteDocument("properties", tenant.id)
